fix(ci): line-limit guard — split oversized unit test modules
